attempt to fix: ValueError("invalid literal for int() with base 10: 'b'")
The encoding sometimes overflows the 8 bits allocated, and sometimes the
encoding function returns an int that is interpreted by Python as a negative
number.

I'm at a loss when it comes to the latter problem; my understanding is that for
8-bit ints, this should never happen; 2's complement interpretation doesn't
happen in eight bits in Python (but apparently it does).

* rigorously clip the constituent components of the encoding
* clip the constant to -10..-0.001,0,0.001..10