fix: increase the exponent bias so that the encoded number is non-negative
This should actually fix the encoding issue.

What seems to have happened is that we increased the precision from 0.01 to
0.001, but never adjusted the exponent bias. As a result, the biased exponent
was still negative (the biasing is a normalization to the non-negative range,
so this ought never to happen), and thereby the encoded number was still
negative (because (-n) << m | x == - ((n << m) | x), and is negative for any
non-negative n, m, x).

We increase the bias, so that for our range below 1, all the way to 0.001, the
biased exponent is >= 0:

```
>>> import numpy as np
>>> np.log2(0.001)
-9.965784284662087
>>> np.floor(_) + 10
0
```